seller profile contact module

+++ src/controllers/app/seller/profile.controller.js
@@ -110,7 +110,6 @@ export const updateCompanyInfo = asyncHandler ( async( req, resp) => {
 export const updateContactPerson = asyncHandler ( async( req, resp)=>{
     const { company_name, gst_no } = req.body;    
     const loggedUser = req.user._id;
-
     
     const schema = Joi.object({
         company_name: Joi.string()
@@ -131,44 +130,31 @@ export const updateContactPerson = asyncHandler ( async( req, resp)=>{
     });
     await schema.validateAsync({ company_name, gst_no }, { abortEarly: false });
  
-    const session = await mongoose.startSession();
-     try {
-        session.startTransaction();
+    const user = await User.findById(loggedUser);
+    let businessDetails;
 
-        const user = await User.findById(loggedUser).session(session);
-        let businessDetails;
-
-        if (user.business_details) {
-            // Update existing
-            businessDetails = await BusinessDetails.findById(user.business_details).session(session);
-            businessDetails.company_name = company_name;
-            businessDetails.gst_number = gst_no;
-            await businessDetails.save({ session });
-        } else {
-            // Create new
-            businessDetails = await BusinessDetails.create([{ company_name:company_name, gst_number:gst_no,}], { session });
-            user.business_details = businessDetails[0]._id;
-            await user.save({ session });
-        }
-        
-        await session.commitTransaction();
-        session.endSession();
-
-        const updatedUser  = await User.findById(loggedUser).select("-otp -otpExpires -refreshToken -__v -createdAt -updatedAt")
-                        .populate({
-                                path: 'business_details',
-                                select: '-_id -createdAt -updatedAt' 
-                            });
-
-        return resp.status(200).json( new ApiResponse(200, updatedUser , "Company information updated successfully.") );
-    
-    } catch (error) {
-        await session.abortTransaction();
-        session.endSession();
-        throw error;
+    if (user.business_details) {
+        businessDetails = await BusinessDetails.findById(user.business_details);
+        businessDetails.company_name = company_name;
+        businessDetails.gst_number = gst_no;
+        await businessDetails.save({ session });
+    } else {
+        // Create new
+        businessDetails = await BusinessDetails.create([{ company_name:company_name, gst_number:gst_no,}], { session });
+        user.business_details = businessDetails[0]._id;
+        await user.save({ session });
     }
+    
+    const updatedUser  = await User.findById(loggedUser).select("-otp -otpExpires -refreshToken -__v -createdAt -updatedAt")
+                        .populate({
+                            path: 'business_details',
+                            select: '-_id -createdAt -updatedAt' 
+                        });
+
+    return resp.status(200).json( new ApiResponse(200, updatedUser , "Company information updated successfully.") );
 
 })
+
 
 
 
@@ -176,14 +162,17 @@ export const updateContactPerson = asyncHandler ( async( req, resp)=>{
 // Seller Contacts Apis 
 export const getAllContacts = asyncHandler(async (req, res) => {
     const loggedUser = req.user._id;
-    const contacts = await UserContacts.find({  user_id : loggedUser }).sort({ sr_no: 1 }); 
+    const contacts = await UserContacts.find({  user_id : loggedUser })
+                        .populate('division', 'name id')
+                        .select('-deleted -createdAt -updatedAt')
+                        .sort({ sr_no: 1 }); 
     return res.status(200).json( new ApiResponse( 200, contacts , "Contact fetched successfully") );
 });
 
 
 export const getContactById = asyncHandler(async (req, res) => { 
     const {id} = req.params;
-    const contacts = await UserContacts.findById(id).select("-deleted");
+    const contacts = await UserContacts.findById(id).populate('division', 'name id').select("-createdAt -updatedAt");
     if (!contacts) {
         return res.status(400).json(new ApiError(400, null, "Contact not found"));
     }
@@ -191,12 +180,9 @@ export const getContactById = asyncHandler(async (req, res) => {
 });
 
 
-
-
 export const createContact = asyncHandler(async (req, res) => {
     const { division, contact_person, address, mobile_no, landline_no, toll_free_no, email, fax_no } = req.body;
-
-    const loggedUser = req.user._id;
+    const loggedUser = req.user.id;
     
     const schema = Joi.object({
       division: Joi.string().min(3).max(100).required().messages({ "string.empty": "Division is required", "string.min": "Division must be at least 3 characters",}),  
@@ -219,37 +205,29 @@ export const createContact = asyncHandler(async (req, res) => {
     });
   
     await schema.validateAsync(req.body, { abortEarly: false });
-  
-    const session = await mongoose.startSession();
-    session.startTransaction();  
-    try {
-      const lastContact = await UserContacts.findOne({ user_id:loggedUser }).sort({ sr_no: -1 });
-      const sr_no = lastContact ? lastContact.sr_no + 1 : 1;
-  
-      const contact = await UserContacts.create({ user_id:loggedUser,  division, contact_person,  address,  mobile_no,  landline_no, toll_free_no,  email,fax_no, sr_no, });
-  
-      await session.commitTransaction();
-      session.endSession();
-  
-      return res.status(201).json(new ApiResponse(201, contact , "Contact added successfully"));
-    } catch (error) {
-      await session.abortTransaction();
-      session.endSession();
-      return res.status(500).json(new ApiResponse(500, "Something went wrong", error.message));
+
+    const isExist = await UserContacts.findOne({ user_id:loggedUser, division:division  });
+    if(isExist){
+        return res.status(400).json(new ApiError(400, null, "Contact already exist on this division"));
     }
-    
+
+    const lastContact = await UserContacts.findOne({ user_id:loggedUser  }).sort({ sr_no: -1 });
+    const sr_no = lastContact ? lastContact.sr_no + 1 : 1;
+
+    const contact = await UserContacts.create({ user_id:loggedUser,  division, contact_person,  address,  mobile_no,  landline_no, toll_free_no,  email,fax_no, sr_no, });
+
+    const populateContact = await UserContacts.findById(contact.id).populate('division', 'name id').select("-createdAt -updatedAt");
+    return res.status(201).json(new ApiResponse(201, populateContact , "Contact added successfully"));
 });
 
 
-
-
 export const updateContact = asyncHandler(async (req, res) => {
-    const loggedUser = req.user._id;
+    const loggedUser = req.user.id;    
     const { id } = req.params;
 
     // Joi schema for partial update (PATCH)
     const schema = Joi.object({
-        division: Joi.string().min(3).max(100).optional(),
+        division: Joi.string(),
         contact_person: Joi.string().min(3).max(100).optional(),
         address: Joi.object({
             address_line: Joi.string().min(5).max(200).optional(),
@@ -269,58 +247,56 @@ export const updateContact = asyncHandler(async (req, res) => {
 
     await schema.validateAsync(req.body, { abortEarly: false });
 
-    try {
-            // Find and update contact
-            const updatedContact = await UserContacts.findOneAndUpdate(
-            { _id: id, user_id: loggedUser }, 
-            { $set: req.body },
-            { new: true, runValidators: true }
-            );
 
-            if (!updatedContact) {
-                return res.status(404).json(new ApiResponse(404, null, "Contact not found"));
-            }
-
-            return res.status(200).json(new ApiResponse(200, updatedContact, "Contact updated successfully"));
-    } catch (error) {
-        return res.status(500).json(new ApiResponse(500, null, error.message));
+     const existingContact = await UserContacts.findOne({
+            user_id: loggedUser, division: req.body.division, _id: { $ne: id },
+    });
+    if (existingContact) {
+        return res.status(400).json(new ApiResponse(400, null, "Contact already exists for this division"));
     }
 
+    const updatedContact = await UserContacts.findOneAndUpdate(
+            { _id: id, user_id: loggedUser },  { $set: req.body }, { new: true, runValidators: true }
+    );
+
+    if (!updatedContact) {
+        return res.status(404).json(new ApiResponse(404, null, "Contact not found"));
+    }
+    const populateContact = await UserContacts.findById(updatedContact.id).populate('division', 'name id').select("-createdAt -updatedAt");
+    return res.status(200).json(new ApiResponse(200, populateContact, "Contact updated successfully"));
+   
 });
+
+
 
 export const deleteContact = asyncHandler(async (req, res) => {
     const { id } = req.params;
-    const loggedUser = req.user._id;
+    const loggedUser = req.user.id;
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(400).json(new ApiResponse(400, null, "Invalid contact id"));
     }
 
-    const contact = await UserContacts.findOneAndDelete({
-        _id: id,
-        user_id: loggedUser, 
-    });
-
-    if (!contact) {
-        return res.status(404).json(new ApiResponse(404, null, "Contact not found"));
-    }
+    const contact = await UserContacts.findOneAndDelete({ _id: id, user_id: loggedUser });
+    if (!contact) { return res.status(404).json(new ApiResponse(404, null, "Contact not found")) }
 
     return res.status(200).json(new ApiResponse(200, contact, "Contact deleted successfully"));
+
 });
 
 
 export const reorderContacts = asyncHandler( async ( req, res) =>{
-    const { testimonials }  = req.body
-    if (!Array.isArray(testimonials)) {
-            return res.status(400).json(new ApiResponse(400, null, "Invalid testimonials payload"));
+    const { contacts }  = req.body
+    if (!Array.isArray(contacts)) {
+            return res.status(400).json(new ApiResponse(400, null, "Invalid contacts payload"));
     }
-    for (const cat of testimonials) {        
+    for (const cat of contacts) {        
         if (!cat.id || typeof cat.sr_no !== 'number') continue;
-        await Testimonials.findByIdAndUpdate(cat.id, { sr_no: cat.sr_no});
+        await UserContacts.findByIdAndUpdate(cat.id, { sr_no: cat.sr_no});
     }
 
-    const newTestimonials = await Testimonials.find({ deleted: { $ne:true } }).sort({ sr_no: 1 });
-    return res.status(200).json(new ApiResponse(200, newTestimonials, "Sr numbers updated"));
+    const newContacts = await UserContacts.find().sort({ sr_no: 1 });
+    return res.status(200).json(new ApiResponse(200, newContacts, "Contacts numbers updated"));
 
 })
